+++ libs/docs/platform/table/e2e/table.po.ts
@@ -73,7 +73,7 @@ export class TablePo extends PlatformBaseComponentPo {
     optionMultiple = this.playgroundSelectionModeDropdown + ' option[value="multiple"]';
     tableCellClass = '.fd-table__cell';
     playgroundSchemaInput = '.form-control.fd-input';
-    toolbarText = '.fd-label.fd-toolbar__overflow-label';
+    toolbarText = '.fd-toolbar__title';
     dropdownList = '.fd-select-options';
     dropdownOption = 'li[fd-option].fd-list__item ';
     dialogButton = 'fd-dialog-body .sort-row__actions .fd-button';